refactor(profile): modularize 1574-line ProfileView into focused components
- Extract ProfileShareButtons: share buttons (X, Facebook, Reddit, copy link)
- Extract ProfileReposSection: sortable repo grid with lang colors
- Extract ProfileBadgeModal: native dialog for add/remove program badges
- Extract ProfileDownloadCard: hidden card rendered to PNG via html-to-image
- Update ProfileView to import and compose extracted components
- Remove 5 unused sub-components (ProfileHeader, StatsGrid, TopRepos,
  TechStack, Organizations) that were never imported anywhere
- ProfileView reduced from 1574 to ~480 lines
