+++ src/components/profile/ProfileView.tsx
@@ -12,6 +12,9 @@ import type { ContributorStats, Org, TechEntry, HeatmapWeek, BadgeItem } from "@
 import { toPng } from "html-to-image";
 import { supabase } from "@/lib/supabase";
 import { LANG_COLORS } from "@/lib/languages";
+import { ProfileShareButtons } from "@/components/profile/ProfileShareButtons";
+import { ProfileReposSection } from "@/components/profile/ProfileReposSection";
+import { ProfileBadgeModal } from "@/components/profile/ProfileBadgeModal";
 
 interface GitHubUser {
   login: string;
@@ -81,7 +84,6 @@ interface ProfileExtras {
   currentStreak: number;
   longestStreak: number;
   score: number;
-  /** ISO 8601 timestamp from Supabase profiles.updated_at — null if the user has never synced. */
   updatedAt: string | null;
   badges: BadgeItem[];
   profileId: string | null;
@@ -89,7 +91,6 @@ interface ProfileExtras {
   mergedPRs: MergedPR[];
 }
 
-/** Format an ISO timestamp as a human-readable relative string for the profile header. */
 function formatUpdatedAt(iso: string): string {
   const diffMs = Date.now() - new Date(iso).getTime();
   const diffDays = Math.floor(diffMs / (1000 * 60 * 60 * 24));
@@ -180,6 +181,162 @@ function ProfileFreshness({ username, updatedAt }: { username: string; updatedAt
   );
 }
 
+function ProfileDownloadCard({
+  user,
+  stats,
+  score,
+  displayName,
+}: {
+  user: GitHubUser;
+  stats: ContributorStats;
+  score: number;
+  displayName: string;
+}) {
+  const cardRef = useRef<HTMLDivElement>(null);
+  const [isDownloading, setIsDownloading] = useState(false);
+
+  const handleDownload = async () => {
+    if (!cardRef.current) return;
+    setIsDownloading(true);
+    try {
+      await new Promise((resolve) => setTimeout(resolve, 100));
+      const dataUrl = await toPng(cardRef.current, {
+        cacheBust: true,
+        pixelRatio: 2,
+        style: { transform: "scale(1)", transformOrigin: "top left" },
+      });
+      const link = document.createElement("a");
+      link.download = `${user.login}-ossfolio-card.png`;
+      link.href = dataUrl;
+      link.click();
+    } catch (err) {
+      console.error("Failed to download profile card:", err);
+    } finally {
+      setIsDownloading(false);
+    }
+  };
+
+  return (
+    <>
+      <button
+        type="button"
+        onClick={handleDownload}
+        disabled={isDownloading}
+        style={{
+          display: "inline-flex",
+          alignItems: "center",
+          gap: "6px",
+          padding: "7px 14px",
+          fontSize: "13px",
+          fontWeight: 500,
+          color: isDownloading ? "var(--color-ink-mute)" : "var(--color-ink)",
+          backgroundColor: "var(--color-canvas-soft)",
+          border: "1px solid var(--color-hairline-strong)",
+          borderRadius: "6px",
+          cursor: isDownloading ? "not-allowed" : "pointer",
+          lineHeight: 1,
+        }}
+        aria-label="Download profile card as PNG"
+      >
+        {isDownloading ? (
+          <>
+            <svg className="animate-spin" width="13" height="13" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" aria-hidden="true">
+              <circle cx="12" cy="12" r="10" stroke="var(--color-hairline-strong)" strokeWidth="4" style={{ opacity: 0.25 }} />
+              <path fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z" />
+            </svg>
+            Generating...
+          </>
+        ) : (
+          <>
+            <svg width="13" height="13" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" aria-hidden="true">
+              <path d="M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4" />
+              <polyline points="7 10 12 15 17 10" />
+              <line x1="12" y1="15" x2="12" y2="3" />
+            </svg>
+            Download card
+          </>
+        )}
+      </button>
+
+      <div style={{ position: "fixed", left: "-9999px", top: "-9999px", overflow: "hidden", pointerEvents: "none" }}>
+        <div
+          ref={cardRef}
+          style={{
+            width: "600px",
+            height: "300px",
+            padding: "32px",
+            backgroundColor: "#1c1c1c",
+            border: "1px solid rgba(255, 255, 255, 0.1)",
+            borderRadius: "12px",
+            display: "flex",
+            flexDirection: "column",
+            justifyContent: "space-between",
+            boxSizing: "border-box",
+            fontFamily: "system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif",
+          }}
+        >
+          <div style={{ display: "flex", justifyContent: "space-between", alignItems: "flex-start", gap: "24px" }}>
+            <div style={{ display: "flex", flexDirection: "column", gap: "24px" }}>
+              <div style={{ display: "flex", gap: "16px", alignItems: "center" }}>
+                <img
+                  src={user.avatar_url}
+                  alt={displayName}
+                  style={{ width: "64px", height: "64px", borderRadius: "9999px", border: "1px solid rgba(255, 255, 255, 0.15)", objectFit: "cover" }}
+                  crossOrigin="anonymous"
+                />
+                <div>
+                  <div style={{ fontSize: "18px", fontWeight: 600, color: "#ffffff", letterSpacing: "-0.3px", lineHeight: 1.2 }}>
+                    {displayName}
+                  </div>
+                  <div style={{ fontSize: "13px", color: "#9a9a9a", marginTop: "2px" }}>
+                    @{user.login}
+                  </div>
+                </div>
+              </div>
+              <div>
+                <div style={{ fontSize: "10px", textTransform: "uppercase", letterSpacing: "1px", color: "#9a9a9a", fontWeight: 600 }}>
+                  Contributor Score
+                </div>
+                <div style={{ fontSize: "44px", fontWeight: 700, color: "#3ecf8e", marginTop: "4px", lineHeight: 1 }}>
+                  {score}
+                </div>
+              </div>
+            </div>
+
+            <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", gap: "12px", width: "260px" }}>
+              {[
+                { label: "Commits", value: stats.totalCommits },
+                { label: "PRs", value: stats.totalPRs },
+                { label: "Issues", value: stats.totalIssues },
+                { label: "Reviews", value: stats.totalReviews },
+              ].map((stat) => (
+                <div key={stat.label} style={{ border: "1px solid rgba(255, 255, 255, 0.08)", borderRadius: "8px", padding: "12px 14px", backgroundColor: "#202020", display: "flex", flexDirection: "column", justifyContent: "center" }}>
+                  <div style={{ fontSize: "20px", fontWeight: 600, color: "#ffffff", lineHeight: 1.1 }}>
+                    {stat.value.toLocaleString("en-US")}
+                  </div>
+                  <div style={{ fontSize: "11px", color: "#9a9a9a", marginTop: "4px", fontWeight: 500 }}>
+                    {stat.label}
+                  </div>
+                </div>
+              ))}
+            </div>
+          </div>
+
+          <div style={{ display: "flex", justifyContent: "space-between", alignItems: "center", borderTop: "1px solid rgba(255, 255, 255, 0.08)", paddingTop: "16px", marginTop: "16px" }}>
+            <div style={{ display: "flex", alignItems: "center", gap: "6px" }}>
+              <span style={{ width: "8px", height: "8px", borderRadius: "50%", backgroundColor: "#3ecf8e" }} />
+              <span style={{ fontSize: "14px", fontWeight: 600, color: "#ffffff", letterSpacing: "-0.2px" }}>OSSfolio</span>
+            </div>
+            <span style={{ fontSize: "11px", fontFamily: "ui-monospace, Menlo, Monaco, Consolas, monospace", color: "#707070" }}>
+              ossfolio.qzz.io
+            </span>
+          </div>
+        </div>
+      </div>
+    </>
+  );
+}
+
 export function ProfileView({
   user,
   repos,
@@ -196,18 +353,14 @@ export function ProfileView({
   rateLimited,
   mergedPRs,
 }: { user: GitHubUser; repos: GitHubRepo[] } & ProfileExtras & { rateLimited?: boolean }) {
-  const [copied, setCopied] = useState(false);
-  const [repoSort, setRepoSort] = useState<"stars" | "forks" | "updated">("stars");
-  const [isDownloading, setIsDownloading] = useState(false);
   const searchRef = useRef<HTMLInputElement>(null);
   const [repoFilter, setRepoFilter] = useState("");
 
   const focusSearch = useCallback(() => searchRef.current?.focus(), []);
   useKeyboardShortcuts({ onSlash: focusSearch });
-  const cardRef = useRef<HTMLDivElement>(null);
+
   const [showBackToTop, setShowBackToTop] = useState(false);
 
-  // Program Badges State
   const sanitizeBadges = (raw: any[]): BadgeItem[] => {
     if (!Array.isArray(raw)) return [];
     return raw
@@ -229,11 +382,6 @@ export function ProfileView({
   const [badgesList, setBadgesList] = useState<BadgeItem[]>(() => sanitizeBadges(badges));
   const [authUser, setAuthUser] = useState<any>(null);
   const [isBadgeModalOpen, setIsBadgeModalOpen] = useState(false);
-  const [selectedProgram, setSelectedProgram] = useState("GSSoC");
-  const currentYear = new Date().getFullYear();
-  const [selectedYear, setSelectedYear] = useState(currentYear);
-  const [isSavingBadge, setIsSavingBadge] = useState(false);
-  const dialogRef = useRef<HTMLDialogElement>(null);
 
   const isOwner = !!(
     authUser && (
@@ -242,13 +390,10 @@ export function ProfileView({
     )
   );
 
-  // Sync badges from props in effect (instead of setState during render)
   useEffect(() => {
-    // eslint-disable-next-line react-hooks/set-state-in-effect
     setBadgesList(sanitizeBadges(badges));
   }, [badges]);
 
-  // Scroll visibility effect
   useEffect(() => {
     const onScroll = () => setShowBackToTop(window.scrollY > 400);
     onScroll();
@@ -256,14 +401,13 @@ export function ProfileView({
     return () => window.removeEventListener("scroll", onScroll);
   }, []);
 
-  // Resolve the current session on mount and keep it in sync with auth changes
   useEffect(() => {
     let active = true;
     supabase.auth.getSession().then(({ data }) => {
       if (active) setAuthUser(data.session?.user ?? null);
     });
     const { data: sub } = supabase.auth.onAuthStateChange((_event, session) => {
-      if (active) setAuthUser(session?.user ?? null);
+      setAuthUser(session?.user ?? null);
     });
     return () => {
       active = false;
@@ -271,44 +415,6 @@ export function ProfileView({
     };
   }, []);
 
-  // Sync native dialog element with state and setup event listeners
-  useEffect(() => {
-    const dialog = dialogRef.current;
-    if (!dialog) return;
-
-    if (isBadgeModalOpen) {
-      if (!dialog.open) dialog.showModal();
-    } else {
-      if (dialog.open) dialog.close();
-    }
-
-    const handleClose = () => {
-      setIsBadgeModalOpen(false);
-    };
-
-    const handleClick = (e: MouseEvent) => {
-      if (e.target === dialog) {
-        const rect = dialog.getBoundingClientRect();
-        const clickInside =
-          rect.top <= e.clientY &&
-          e.clientY <= rect.top + rect.height &&
-          rect.left <= e.clientX &&
-          e.clientX <= rect.left + rect.width;
-        if (!clickInside) {
-          setIsBadgeModalOpen(false);
-        }
-      }
-    };
-
-    dialog.addEventListener("close", handleClose);
-    dialog.addEventListener("click", handleClick);
-    return () => {
-      dialog.removeEventListener("close", handleClose);
-      dialog.removeEventListener("click", handleClick);
-    };
-  }, [isBadgeModalOpen, isOwner]);
-
-  // Conditional rendering
   if (rateLimited) {
     return (
       <div style={{ color: 'var(--color-ink-mute)', backgroundColor: 'var(--color-canvas-soft)', padding: '16px', borderRadius: '8px', textAlign: 'center', marginBottom: '24px' }}>
@@ -324,65 +430,11 @@ export function ProfileView({
       : `https://${user.blog}`
     : null;
 
-  const handleAddBadge = async () => {
-    if (!profileId) {
-      alert("Please sync your profile first before adding badges.");
-      return;
-    }
-    const targetProfileId = profileId;
-    setIsSavingBadge(true);
-    try {
-      // Create new list of badges
-      const existingBadgeIndex = badgesList.findIndex(
-        (b) => b.program === selectedProgram
-      );
-      let updatedList: BadgeItem[] = [];
-      if (existingBadgeIndex > -1) {
-        // Append year if not already present, sort desc
-        const currentYears = badgesList[existingBadgeIndex].years;
-        const updatedYears = currentYears.includes(selectedYear)
-          ? currentYears
-          : [...currentYears, selectedYear].sort((a, b) => b - a);
-        updatedList = badgesList.map((b, idx) =>
-          idx === existingBadgeIndex ? { ...b, years: updatedYears } : b
-        );
-      } else {
-        // Add new badge
-        updatedList = [
-          ...badgesList,
-          { program: selectedProgram, years: [selectedYear] },
-        ];
-      }
-
-      const { error } = await supabase
-        .from("profiles")
-        .upsert({
-          id: targetProfileId,
-          username: user.login,
-          badges: updatedList,
-          updated_at: new Date().toISOString(),
-        });
-
-      if (error) {
-        console.error("Failed to save badge to database:", error.message);
-        alert(`Failed to save badge: ${error.message}`);
-      } else {
-        setBadgesList(updatedList);
-        setIsBadgeModalOpen(false);
-      }
-    } catch (err) {
-      console.error("Error saving badge:", err);
-    } finally {
-      setIsSavingBadge(false);
-    }
-  };
-
   const handleRemoveBadge = async (program: string) => {
     if (!profileId) {
       alert("Please sync your profile first before removing badges.");
       return;
     }
-    const targetProfileId = profileId;
     const confirmRemove = confirm(`Are you sure you want to remove the ${program} badge?`);
     if (!confirmRemove) return;
 
@@ -391,14 +443,13 @@ export function ProfileView({
       const { error } = await supabase
         .from("profiles")
         .upsert({
-          id: targetProfileId,
+          id: profileId,
           username: user.login,
           badges: updatedList,
           updated_at: new Date().toISOString(),
         });
 
       if (error) {
-        console.error("Failed to remove badge from database:", error.message);
         alert(`Failed to remove badge: ${error.message}`);
       } else {
         setBadgesList(updatedList);
@@ -408,58 +459,14 @@ export function ProfileView({
     }
   };
 
-  const handleCopyLink = async () => {
-    try {
-      await navigator.clipboard.writeText(window.location.href);
-      setCopied(true);
-      setTimeout(() => setCopied(false), 2000);
-    } catch (err) {
-      console.error("Copy to clipboard failed:", err);
-    }
-  };
-
-  const handleDownloadCard = async () => {
-    if (!cardRef.current) return;
-    setIsDownloading(true);
-    try {
-      // Small delay to ensure images have finished rendering
-      await new Promise((resolve) => setTimeout(resolve, 100));
-
-      const dataUrl = await toPng(cardRef.current, {
-        cacheBust: true,
-        pixelRatio: 2,
-        style: {
-          transform: "scale(1)",
-          transformOrigin: "top left",
-        },
-      });
-      const link = document.createElement("a");
-      link.download = `${user.login}-ossfolio-card.png`;
-      link.href = dataUrl;
-      link.click();
-    } catch (err) {
-      console.error("Failed to download profile card:", err);
-    } finally {
-      setIsDownloading(false);
-    }
-  };
-
-  // Show a "Back to top" button once the visitor scrolls past 400px.
-  
-
-
   const scrollToTop = () =>
     window.scrollTo({ top: 0, behavior: "smooth" });
 
-  // Total stars across the repos shown on this page
   const totalStars = repos.reduce((sum, r) => sum + (r.stargazers_count ?? 0), 0);
-
   const totalForks = repos.reduce((sum, r) => sum + (r.forks_count ?? 0), 0);
 
   return (
     <div style={{ maxWidth: "56rem", margin: "0 auto", padding: "48px 20px 80px" }}>
-
-
       {/* Profile header */}
       <div style={{ display: "flex", alignItems: "flex-start", gap: "24px", flexWrap: "wrap", paddingBottom: "40px", borderBottom: "1px solid var(--color-hairline)" }}>
         <Image
@@ -482,10 +489,8 @@ export function ProfileView({
             </p>
           )}
 
-          {/* Profile Freshness Indicator */}
           <ProfileFreshness username={user.login} updatedAt={updatedAt ?? undefined} />
 
-          {/* Dynamic Contributor Tier Badge */}
           {(() => {
             let tierName = "Bronze Contributor";
             let tierColor = "#cd7f32";
@@ -509,20 +514,7 @@ export function ProfileView({
             }
             return (
               <div style={{ marginTop: "12px", display: "inline-flex", alignItems: "center" }}>
-                <span
-                  style={{
-                    fontSize: "11px",
-                    fontWeight: 700,
-                    textTransform: "uppercase",
-                    letterSpacing: "0.05em",
-                    color: tierColor,
-                    backgroundColor: tierBg,
-                    border: `1px solid ${tierColor}`,
-                    padding: "3px 8px",
-                    borderRadius: "4px",
-                    boxShadow: "0 2px 8px rgba(0,0,0,0.05)"
-                  }}
-                >
+                <span style={{ fontSize: "11px", fontWeight: 700, textTransform: "uppercase", letterSpacing: "0.05em", color: tierColor, backgroundColor: tierBg, border: `1px solid ${tierColor}`, padding: "3px 8px", borderRadius: "4px", boxShadow: "0 2px 8px rgba(0,0,0,0.05)" }}>
                   {tierName}
                 </span>
               </div>
@@ -545,8 +537,6 @@ export function ProfileView({
                 rel="noopener noreferrer"
                 aria-label={`Personal website ${website.replace(/^https?:\/\//, "")} (opens in a new tab)`}
                 style={{ fontSize: "13px", color: "var(--color-ink-mute)", display: "flex", alignItems: "center", gap: "5px", textDecoration: "none" }}
-                onMouseEnter={(e) => (e.currentTarget.style.color = "var(--color-ink)")}
-                onMouseLeave={(e) => (e.currentTarget.style.color = "var(--color-ink-mute)")}
               >
                 <svg width="13" height="13" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
                   <path d="M10 13a5 5 0 0 0 7.54.54l3-3a5 5 0 0 0-7.07-7.07l-1.72 1.71" />
@@ -562,8 +552,6 @@ export function ProfileView({
                 rel="noopener noreferrer"
                 aria-label={`Twitter profile of @${user.twitter_username} (opens in a new tab)`}
                 style={{ fontSize: "13px", color: "var(--color-ink-mute)", display: "flex", alignItems: "center", gap: "5px", textDecoration: "none" }}
-                onMouseEnter={(e) => (e.currentTarget.style.color = "var(--color-ink)")}
-                onMouseLeave={(e) => (e.currentTarget.style.color = "var(--color-ink-mute)")}
               >
                 <svg aria-hidden="true" width="13" height="13" viewBox="0 0 24 24" fill="currentColor">
                   <path d="M18.244 2.25h3.308l-7.227 8.26 8.502 11.24H16.17l-5.214-6.817L4.99 21.75H1.68l7.73-8.835L1.254 2.25H8.08l4.713 6.231zm-1.161 17.52h1.833L7.084 4.126H5.117z" />
@@ -577,8 +565,6 @@ export function ProfileView({
               rel="noopener noreferrer"
               aria-label={`GitHub profile of ${displayName} (opens in a new tab)`}
               style={{ fontSize: "13px", color: "var(--color-ink-mute)", display: "flex", alignItems: "center", gap: "5px", textDecoration: "none" }}
-              onMouseEnter={(e) => (e.currentTarget.style.color = "var(--color-ink)")}
-              onMouseLeave={(e) => (e.currentTarget.style.color = "var(--color-ink-mute)")}
             >
               <svg aria-hidden="true" width="13" height="13" viewBox="0 0 24 24" fill="currentColor">
                 <path d="M12 0C5.37 0 0 5.37 0 12c0 5.31 3.435 9.795 8.205 11.385.6.105.825-.255.825-.57 0-.285-.015-1.23-.015-2.235-3.015.555-3.795-.735-4.035-1.41-.135-.345-.72-1.41-1.23-1.695-.42-.225-1.02-.78-.015-.795.945-.015 1.62.87 1.845 1.23 1.08 1.815 2.805 1.305 3.495.99.105-.78.42-1.305.765-1.605-2.67-.3-5.46-1.335-5.46-5.925 0-1.305.465-2.385 1.23-3.225-.12-.3-.54-1.53.12-3.18 0 0 1.005-.315 3.3 1.23.96-.27 1.98-.405 3-.405s2.04.135 3 .405c2.295-1.56 3.3-1.23 3.3-1.23.66 1.65.24 2.88.12 3.18.765.84 1.23 1.905 1.23 3.225 0 4.605-2.805 5.625-5.475 5.925.435.375.81 1.095.81 2.22 0 1.605-.015 2.895-.015 3.3 0 .315.225.69.825.57A12.02 12.02 0 0 0 24 12c0-6.63-5.37-12-12-12z" />
@@ -587,226 +573,29 @@ export function ProfileView({
             </a>
           </div>
 
-          {/* Action buttons */}
+          <div style={{ marginTop: "14px" }}>
+            <ProfileShareButtons username={user.login} score={score} />
+          </div>
+
           <div style={{ marginTop: "14px", display: "flex", gap: "8px", flexWrap: "wrap" }}>
-            <button
-              type="button"
-              onClick={() => {
-                const profileUrl = `https://ossfolio.qzz.io/${user.login}`;
-                const text = `My open source contributor score is ${score} on OSSfolio: ${profileUrl} #opensource`;
-                window.open(
-                  `https://twitter.com/intent/tweet?text=${encodeURIComponent(text)}`,
-                  "_blank",
-                  "noopener,noreferrer"
-                );
-              }}
+            <ProfileDownloadCard user={user} stats={stats} score={score} displayName={displayName} />
+            <input
+              ref={searchRef}
+              type="text"
+              placeholder="Search repositories..."
+              value={repoFilter}
+              onChange={(e) => setRepoFilter(e.target.value)}
               style={{
-                display: "inline-flex",
-                alignItems: "center",
-                gap: "6px",
                 padding: "7px 14px",
                 fontSize: "13px",
-                fontWeight: 500,
+                border: "1px solid var(--color-hairline-strong)",
+                borderRadius: "6px",
+                backgroundColor: "var(--color-canvas-soft)",
                 color: "var(--color-ink)",
-                backgroundColor: "var(--color-canvas-soft)",
-                border: "1px solid var(--color-hairline-strong)",
-                borderRadius: "6px",
-                cursor: "pointer",
-                lineHeight: 1,
+                outline: "none",
               }}
-              onMouseEnter={(e) => {
-                e.currentTarget.style.borderColor = "var(--color-ink)";
-                e.currentTarget.style.backgroundColor = "var(--color-hairline)";
-              }}
-              onMouseLeave={(e) => {
-                e.currentTarget.style.borderColor = "var(--color-hairline-strong)";
-                e.currentTarget.style.backgroundColor = "var(--color-canvas-soft)";
-              }}
-              aria-label="Share profile on X (Twitter)"
-            >
-              <svg width="13" height="13" viewBox="0 0 24 24" fill="currentColor" aria-hidden="true">
-                <path d="M18.244 2.25h3.308l-7.227 8.26 8.502 11.24H16.17l-5.214-6.817L4.99 21.75H1.68l7.73-8.835L1.254 2.25H8.08l4.713 6.231zm-1.161 17.52h1.833L7.084 4.126H5.117z" />
-              </svg>
-              Share on X
-            </button>
-            {/* Facebook Share Button */}
-            <button
-              type="button"
-              onClick={() => {
-                const profileUrl = `https://ossfolio.qzz.io/${user.login}`;
-                const fbUrl = `https://www.facebook.com/sharer/sharer.php?u=${encodeURIComponent(profileUrl)}`;
-                window.open(fbUrl, "_blank", "noopener,noreferrer");
-              }}
-              style={{
-                display: "inline-flex",
-                alignItems: "center",
-                gap: "6px",
-                padding: "7px 14px",
-                fontSize: "13px",
-                fontWeight: 500,
-                color: "var(--color-ink)",
-                backgroundColor: "var(--color-canvas-soft)",
-                border: "1px solid var(--color-hairline-strong)",
-                borderRadius: "6px",
-                cursor: "pointer",
-                lineHeight: 1,
-              }}
-              onMouseEnter={(e) => {
-                e.currentTarget.style.borderColor = "var(--color-ink)";
-                e.currentTarget.style.backgroundColor = "var(--color-hairline)";
-              }}
-              onMouseLeave={(e) => {
-                e.currentTarget.style.borderColor = "var(--color-hairline-strong)";
-                e.currentTarget.style.backgroundColor = "var(--color-canvas-soft)";
-              }}
-              aria-label="Share profile on Facebook"
-            >
-              <svg width="13" height="13" viewBox="0 0 24 24" fill="currentColor" aria-hidden="true">
-                <path d="M22 12c0-5.52-4.48-10-10-10S2 6.48 2 12c0 4.99 3.66 9.12 8.44 9.88v-6.99h-2.54V12h2.54V9.69c0-2.5 1.5-3.89 3.8-3.89 1.1 0 2.24.2 2.24.2v2.47h-1.26c-1.24 0-1.63.77-1.63 1.56V12h2.78l-.44 2.89h-2.34v6.99C18.34 21.12 22 16.99 22 12z" />
-              </svg>
-              Share on Facebook
-            </button>
-
-            {/* Reddit Share Button */}
-            <button
-              type="button"
-              onClick={() => {
-                const profileUrl = `https://ossfolio.qzz.io/${user.login}`;
-                const title = `My open source score on OSSfolio`;
-                const redditUrl = `https://www.reddit.com/submit?url=${encodeURIComponent(profileUrl)}&title=${encodeURIComponent(title)}`;
-                window.open(redditUrl, "_blank", "noopener,noreferrer");
-              }}
-              style={{
-                display: "inline-flex",
-                alignItems: "center",
-                gap: "6px",
-                padding: "7px 14px",
-                fontSize: "13px",
-                fontWeight: 500,
-                color: "var(--color-ink)",
-                backgroundColor: "var(--color-canvas-soft)",
-                border: "1px solid var(--color-hairline-strong)",
-                borderRadius: "6px",
-                cursor: "pointer",
-                lineHeight: 1,
-              }}
-              onMouseEnter={(e) => {
-                e.currentTarget.style.borderColor = "var(--color-ink)";
-                e.currentTarget.style.backgroundColor = "var(--color-hairline)";
-              }}
-              onMouseLeave={(e) => {
-                e.currentTarget.style.borderColor = "var(--color-hairline-strong)";
-                e.currentTarget.style.backgroundColor = "var(--color-canvas-soft)";
-              }}
-              aria-label="Share profile on Reddit"
-            >
-              <svg width="13" height="13" viewBox="0 0 24 24" fill="currentColor" aria-hidden="true">
-                <path d="M6.167 8a.83.83 0 0 0-.83.83c0 .459.372.84.83.831a.831.831 0 0 0 0-1.661m1.843 3.647c.315 0 1.403-.038 1.976-.611a.23.23 0 0 0 0-.306.213.213 0 0 0-.306 0c-.353.363-1.126.487-1.67.487-.545 0-1.308-.124-1.671-.487a.213.213 0 0 0-.306 0 .213.213 0 0 0 0 .306c.564.563 1.652.61 1.977.61zm.992-2.807c0 .458.373.83.831.83s.83-.381.83-.83a.831.831 0 0 0-1.66 0z" />
-              </svg>
-              Share on Reddit
-            </button>
-
-            <button
-              type="button"
-              onClick={handleCopyLink}
-              style={{
-                display: "inline-flex",
-                alignItems: "center",
-                gap: "6px",
-                padding: "7px 14px",
-                fontSize: "13px",
-                fontWeight: 500,
-                color: copied ? "#3ecf8e" : "var(--color-ink)",
-                backgroundColor: "var(--color-canvas-soft)",
-                border: `1px solid ${copied ? "#3ecf8e" : "var(--color-hairline-strong)"}`,
-                borderRadius: "6px",
-                cursor: "pointer",
-                lineHeight: 1,
-                transition: "border-color 0.15s, color 0.15s",
-              }}
-              onMouseEnter={(e) => {
-                if (!copied) {
-                  e.currentTarget.style.borderColor = "var(--color-ink)";
-                  e.currentTarget.style.backgroundColor = "var(--color-hairline)";
-                }
-              }}
-              onMouseLeave={(e) => {
-                if (!copied) {
-                  e.currentTarget.style.borderColor = "var(--color-hairline-strong)";
-                  e.currentTarget.style.backgroundColor = "var(--color-canvas-soft)";
-                }
-              }}
-              aria-label="Copy profile link to clipboard"
-            >
-              {copied ? (
-                <>
-                  <svg width="13" height="13" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.5" aria-hidden="true">
-                    <polyline points="20 6 9 17 4 12" />
-                  </svg>
-                  Copied!
-                </>
-              ) : (
-                <>
-                  <svg width="13" height="13" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" aria-hidden="true">
-                    <rect x="9" y="9" width="13" height="13" rx="2" ry="2" />
-                    <path d="M5 15H4a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h9a2 2 0 0 1 2 2v1" />
-                  </svg>
-                  Copy link
-                </>
-              )}
-            </button>
-
-            <button
-              type="button"
-              onClick={handleDownloadCard}
-              disabled={isDownloading}
-              style={{
-                display: "inline-flex",
-                alignItems: "center",
-                gap: "6px",
-                padding: "7px 14px",
-                fontSize: "13px",
-                fontWeight: 500,
-                color: isDownloading ? "var(--color-ink-mute)" : "var(--color-ink)",
-                backgroundColor: "var(--color-canvas-soft)",
-                border: "1px solid var(--color-hairline-strong)",
-                borderRadius: "6px",
-                cursor: isDownloading ? "not-allowed" : "pointer",
-                lineHeight: 1,
-              }}
-              onMouseEnter={(e) => {
-                if (!isDownloading) {
-                  e.currentTarget.style.borderColor = "var(--color-ink)";
-                  e.currentTarget.style.backgroundColor = "var(--color-hairline)";
-                }
-              }}
-              onMouseLeave={(e) => {
-                if (!isDownloading) {
-                  e.currentTarget.style.borderColor = "var(--color-hairline-strong)";
-                  e.currentTarget.style.backgroundColor = "var(--color-canvas-soft)";
-                }
-              }}
-              aria-label="Download profile card as PNG"
-            >
-              {isDownloading ? (
-                <>
-                  <svg className="animate-spin" width="13" height="13" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" aria-hidden="true">
-                    <circle cx="12" cy="12" r="10" stroke="var(--color-hairline-strong)" strokeWidth="4" style={{ opacity: 0.25 }} />
-                    <path fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z" />
-                  </svg>
-                  Generating...
-                </>
-              ) : (
-                <>
-                  <svg width="13" height="13" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" aria-hidden="true">
-                    <path d="M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4" />
-                    <polyline points="7 10 12 15 17 10" />
-                    <line x1="12" y1="15" x2="12" y2="3" />
-                  </svg>
-                  Download card
-                </>
-              )}
-            </button>
+              aria-label="Search repositories"
+            />
           </div>
 
           <div style={{ display: "flex", gap: "20px", marginTop: "14px" }}>
@@ -820,7 +609,6 @@ export function ProfileView({
               <strong style={{ color: "var(--color-ink)", fontWeight: 600 }}>{user.public_repos}</strong> repos
             </span>
           </div>
-
         </div>
       </div>
 
@@ -850,8 +638,6 @@ export function ProfileView({
                   lineHeight: 1,
                   transition: "background-color 0.15s",
                 }}
-                onMouseEnter={(e) => (e.currentTarget.style.backgroundColor = "#24b47e")}
-                onMouseLeave={(e) => (e.currentTarget.style.backgroundColor = "#3ecf8e")}
               >
                 <svg width="14" height="14" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.5">
                   <line x1="12" y1="5" x2="12" y2="19" /><line x1="5" y1="12" x2="19" y2="12" />
@@ -885,48 +671,33 @@ export function ProfileView({
                       fontSize: "13px",
                       fontWeight: 600,
                       boxShadow: "0 2px 4px rgba(0,0,0,0.08)",
-                      transition: "transform 0.15s",
                     }}
                   >
                     <span>{badge.program}</span>
-                    <span
-                      style={{
-                        backgroundColor: "rgba(255, 255, 255, 0.25)",
-                        padding: "2px 6px",
-                        borderRadius: "9999px",
-                        fontSize: "11px",
-                        fontWeight: 500,
-                      }}
-                    >
+                    <span style={{ backgroundColor: "rgba(255, 255, 255, 0.25)", padding: "2px 6px", borderRadius: "9999px", fontSize: "11px", fontWeight: 500 }}>
                       {badge.years.join(", ")}
                     </span>
                     {isOwner && (
-                      <>
-                        <button
-                          type="button"
-                          onClick={() => handleRemoveBadge(badge.program)}
-                          title={`Remove ${badge.program} badge`}
-                          aria-label={`Remove ${badge.program} badge`}
-                          style={{
-                            background: "none",
-                            border: "none",
-                            color: "rgba(255, 255, 255, 0.8)",
-                            cursor: "pointer",
-                            padding: 0,
-                            display: "flex",
-                            alignItems: "center",
-                            fontSize: "16px",
-                            marginLeft: "4px",
-                            lineHeight: 1,
-                          }}
-                          onMouseEnter={(e) => (e.currentTarget.style.color = "#ffffff")}
-                          onMouseLeave={(e) => (e.currentTarget.style.color = "rgba(255, 255, 255, 0.8)")}
-                        >
-                          &times;
-                        </button>
-                        {/* Latest Merged PRs */}
-                        <LatestMergedPRs mergedPRs={mergedPRs} />
-                      </>
+                      <button
+                        type="button"
+                        onClick={() => handleRemoveBadge(badge.program)}
+                        title={`Remove ${badge.program} badge`}
+                        aria-label={`Remove ${badge.program} badge`}
+                        style={{
+                          background: "none",
+                          border: "none",
+                          color: "rgba(255, 255, 255, 0.8)",
+                          cursor: "pointer",
+                          padding: 0,
+                          display: "flex",
+                          alignItems: "center",
+                          fontSize: "16px",
+                          marginLeft: "4px",
+                          lineHeight: 1,
+                        }}
+                      >
+                        &times;
+                      </button>
                     )}
                   </div>
                 );
@@ -936,151 +707,11 @@ export function ProfileView({
         </div>
       )}
 
+      {/* Latest Merged PRs */}
+      <LatestMergedPRs mergedPRs={mergedPRs} />
+
       {/* Repos */}
-      <div style={{ marginTop: "40px" }}>
-        <div style={{ display: "flex", alignItems: "center", justifyContent: "space-between", margin: "0 0 12px 0", flexWrap: "wrap", gap: "12px" }}>
-          <h2 style={{ fontSize: "16px", fontWeight: 600, color: "var(--color-ink)", margin: 0, letterSpacing: "-0.2px" }}>
-            Popular repositories
-          </h2>
-          <div role="group" aria-label="Sort popular repositories" style={{ display: "flex", gap: "6px" }}>
-            {(["stars", "forks", "updated"] as const).map((option) => (
-              <button
-                key={option}
-                type="button"
-                aria-pressed={repoSort === option}
-                onClick={() => setRepoSort(option)}
-                style={{
-                  padding: "4px 10px",
-                  fontSize: "12px",
-                  fontWeight: repoSort === option ? 600 : 400,
-                  color: repoSort === option ? "#171717" : "var(--color-ink-mute)",
-                  backgroundColor: repoSort === option ? "#3ecf8e" : "var(--color-canvas-soft)",
-                  border: repoSort === option ? "none" : "1px solid var(--color-hairline)",
-                  borderRadius: "9999px",
-                  cursor: "pointer",
-                }}
-              >
-                {option === "stars" ? "Stars" : option === "forks" ? "Forks" : "Recent"}
-              </button>
-            ))}
-          </div>
-        </div>
-
-        {repos.length === 0 ? (
-          <div style={{ padding: "40px", border: "1px dashed var(--color-hairline-strong)", borderRadius: "12px", textAlign: "center", backgroundColor: "var(--color-canvas-soft)", margin: "16px 0" }}>
-            <svg style={{ margin: "0 auto 12px", color: "var(--color-ink-mute-2)" }} width="32" height="32" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-              <rect x="2" y="2" width="20" height="20" rx="2" ry="2" />
-              <path d="M12 18H12.01" />
-            </svg>
-            <p style={{ fontSize: "14px", fontWeight: 600, color: "var(--color-ink)", margin: 0 }}>No public repositories found</p>
-            <p style={{ fontSize: "13px", color: "var(--color-ink-mute)", margin: "4px 0 0 0" }}>Create public repositories on GitHub to display them here.</p>
-          </div>
-        ) : (
-          <>
-          <div style={{ display: "grid", gridTemplateColumns: "repeat(auto-fill, minmax(280px, 1fr))", gap: "16px" }}>
-            {[...repos].sort((a, b) => {
-              if (repoSort === "forks") return b.forks_count - a.forks_count;
-              if (repoSort === "updated") return (b.pushed_at || "").localeCompare(a.pushed_at || "");
-              return b.stargazers_count - a.stargazers_count;
-            }).filter((repo) => !repoFilter || repo.name.toLowerCase().includes(repoFilter.toLowerCase()) || (repo.description || "").toLowerCase().includes(repoFilter.toLowerCase())).map((repo) => (
-              <a
-                key={repo.id}
-                href={repo.html_url}
-                target="_blank"
-                rel="noopener noreferrer"
-                style={{
-                  display: "flex",
-                  flexDirection: "column",
-                  gap: "8px",
-                  padding: "20px",
-                  border: "1px solid var(--color-hairline)",
-                  borderRadius: "12px",
-                  textDecoration: "none",
-                  backgroundColor: "var(--color-canvas-soft)",
-                }}
-                onMouseEnter={(e) => {
-                  e.currentTarget.style.borderColor = "var(--color-hairline-strong)";
-                  e.currentTarget.style.boxShadow = "0 1px 3px rgba(0,0,0,0.12)";
-                }}
-                onMouseLeave={(e) => {
-                  e.currentTarget.style.borderColor = "var(--color-hairline)";
-                  e.currentTarget.style.boxShadow = "none";
-                }}
-              >
-                <p style={{ fontSize: "14px", fontWeight: 600, color: "var(--color-ink)", margin: 0, overflow: "hidden", textOverflow: "ellipsis", whiteSpace: "nowrap" }}>
-                  {repo.name}
-                </p>
-                <p style={{ fontSize: "13px", color: "var(--color-ink-mute)", margin: 0, lineHeight: 1.45, display: "-webkit-box", WebkitLineClamp: 2, WebkitBoxOrient: "vertical" as const, overflow: "hidden", minHeight: "38px" }}>
-                  {repo.description || "No description"}
-                </p>
-                {repo.topics && repo.topics.length > 0 && (
-                  <div style={{ display: "flex", flexWrap: "wrap", gap: "4px", marginTop: "8px" }}>
-                    {repo.topics.slice(0, 3).map((topic) => (
-                      <span
-                        key={topic}
-                        style={{
-                          fontSize: "11px",
-                          padding: "2px 8px",
-                          borderRadius: "9999px",
-                          backgroundColor: "var(--color-canvas-soft)",
-                          color: "var(--color-ink-mute)",
-                          border: "1px solid var(--color-hairline)",
-                        }}
-                      >
-                        {topic}
-                      </span>
-                    ))}
-                    {repo.topics.length > 3 && (
-                      <span style={{ fontSize: "11px", padding: "2px 6px", color: "var(--color-ink-mute)" }}>
-                        +{repo.topics.length - 3} more
-                      </span>
-                    )}
-                  </div>
-                )}
-                <div style={{ display: "flex", alignItems: "center", gap: "16px", marginTop: "auto", paddingTop: "8px" }}>
-                  {repo.language && (
-                    <span style={{ display: "flex", alignItems: "center", gap: "5px", fontSize: "12px", color: "var(--color-ink-mute)" }}>
-                      <span style={{ width: "10px", height: "10px", borderRadius: "9999px", backgroundColor: LANG_COLORS[repo.language] ?? "#9a9a9a", flexShrink: 0 }} />
-                      {repo.language}
-                    </span>
-                  )}
-                  <span style={{ display: "flex", alignItems: "center", gap: "4px", fontSize: "12px", color: "var(--color-ink-mute)" }}>
-                    <svg width="13" height="13" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-                      <polygon points="12 2 15.09 8.26 22 9.27 17 14.14 18.18 21.02 12 17.77 5.82 21.02 7 14.14 2 9.27 8.91 8.26 12 2" />
-                    </svg>
-                    {repo.stargazers_count.toLocaleString("en-US")}
-                  </span>
-                  <span style={{ display: "flex", alignItems: "center", gap: "4px", fontSize: "12px", color: "var(--color-ink-mute)" }}>
-                    <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-                      <circle cx="12" cy="18" r="3" /><circle cx="6" cy="6" r="3" /><circle cx="18" cy="6" r="3" />
-                      <path d="M18 9a9 9 0 0 1-9 9M6 9a9 9 0 0 0 9 9" />
-                    </svg>
-                    {repo.forks_count.toLocaleString("en-US")}
-                  </span>
-                </div>
-              </a>
-            ))}
-          </div>
-
-          <div style={{ marginTop: "20px" }}>
-            <a
-              href={`https://github.com/${user.login}?tab=repositories`}
-              target="_blank"
-              rel="noopener noreferrer"
-              aria-label="View all repositories on GitHub (opens in a new tab)"
-              style={{ fontSize: "13px", color: "var(--color-ink-mute)", textDecoration: "none", display: "inline-flex", alignItems: "center", gap: "4px" }}
-              onMouseEnter={(e) => (e.currentTarget.style.color = "var(--color-ink)")}
-              onMouseLeave={(e) => (e.currentTarget.style.color = "var(--color-ink-mute)")}
-            >
-              View all repositories on GitHub
-              <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2">
-                <path d="M5 12h14M12 5l7 7-7 7" />
-              </svg>
-            </a>
-          </div>
-          </>
-        )}
-      </div>
+      <ProfileReposSection repos={repos} username={user.login} />
 
       {/* Contribution stats */}
       <div style={{ marginTop: "44px" }}>
@@ -1211,8 +842,6 @@ export function ProfileView({
                 title={org.name ?? org.login}
                 aria-label={`Organization ${org.name ?? org.login} (opens in a new tab)`}
                 style={{ display: "inline-flex", borderRadius: "8px", overflow: "hidden", border: "1px solid var(--color-hairline)", transition: "border-color 0.15s" }}
-                onMouseEnter={(e) => (e.currentTarget.style.borderColor = "#3ecf8e")}
-                onMouseLeave={(e) => (e.currentTarget.style.borderColor = "var(--color-hairline)")}
               >
                 <Image src={org.avatarUrl} alt={org.login} width={36} height={36} style={{ display: "block" }} />
               </a>
@@ -1222,89 +851,12 @@ export function ProfileView({
       )}
 
       {/* Contribution heatmap with year navigation */}
-      {heatmap.length > 0 && (
-        <div style={{ marginTop: "44px" }}>
-          <h2 style={{ fontSize: "16px", fontWeight: 600, color: "var(--color-ink)", margin: "0 0 16px 0", letterSpacing: "-0.2px" }}>
-            Contribution activity
-          </h2>
-          <div style={{ display: "flex", flexWrap: "wrap", gap: "8px", margin: "0 0 12px 0" }}>
-            {[
-              { label: "Current streak", value: currentStreak },
-              { label: "Longest streak", value: longestStreak },
-            ].map(({ label, value }) => (
-              <span
-                key={label}
-                style={{
-                  display: "inline-flex",
-                  alignItems: "baseline",
-                  gap: "6px",
-                  padding: "6px 12px",
-                  border: "1px solid var(--color-hairline)",
-                  borderRadius: "9999px",
-                  fontSize: "13px",
-                  color: "var(--color-ink-mute)",
-                  backgroundColor: "var(--color-canvas-soft)",
-                }}
-              >
-                <strong style={{ color: "var(--color-ink)", fontWeight: 600 }}>
-                  {value} {value === 1 ? "day" : "days"}
-                </strong>
-                {label}
-              </span>
-            ))}
-          </div>
-          <div
-            style={{
-              overflowX: "auto",
-              padding: "16px",
-              border: "1px solid var(--color-hairline)",
-              borderRadius: "12px",
-              backgroundColor: "var(--color-canvas-soft)",
-            }}
-          >
-            <div style={{ display: "flex", flexDirection: "column", minWidth: "max-content" }}>
-              {/* Month labels */}
-              <div style={{ display: "flex", gap: "3px", marginBottom: "4px", fontSize: "12px", color: "var(--color-ink-mute)" }}>
-                {heatmap.map((week, wi) => {
-                  const month = new Date(week.days[0].date).toLocaleString('en-US', { month: 'short' });
-                  const show = wi === 0 || month !== new Date(heatmap[wi - 1].days[0].date).toLocaleString('en-US', { month: 'short' });
-                  return (
-                    <span key={wi} style={{ width: "11px", textAlign: "center", flexShrink: 0 }}>{show ? month : ""}</span>
-                  );
-                })}
-              </div>
-              {/* Weeks grid */}
-              <div style={{ display: "flex", gap: "3px" }}>
-                {heatmap.map((week, wi) => (
-                  <div key={wi} style={{ display: "flex", flexDirection: "column", gap: "3px" }}>
-                    {week.days.map((day, di) => (
-                      <div
-                        key={di}
-                        title={`${day.count} contributions on ${day.date}`}
-                        style={{ width: "11px", height: "11px", borderRadius: "2px", backgroundColor: day.color, flexShrink: 0 }}
-                      />
-                    ))}
-                  </div>
-                ))}
-              </div>
-            </div>
-          </div>
-          <div style={{ display: "flex", alignItems: "center", justifyContent: "flex-end", gap: "4px", margin: "10px 0 0 0" }}>
-            <span style={{ fontSize: "12px", color: "var(--color-ink-mute)", marginRight: "2px" }}>Less</span>
-            {["var(--color-hairline)", "#9be9a8", "#40c463", "#30a14e", "#216e39"].map((shade) => (
-              <span
-                key={shade}
-                aria-hidden="true"
-                style={{ width: "11px", height: "11px", borderRadius: "2px", backgroundColor: shade.startsWith("var") ? "rgba(128, 128, 128, 0.1)" : shade, flexShrink: 0 }}
-              />
-            ))}
-            <span style={{ fontSize: "12px", color: "var(--color-ink-mute)", marginLeft: "2px" }}>More</span>
-          </div>
-          <p style={{ fontSize: "12px", color: "var(--color-ink-mute)", margin: "10px 0 0 0" }}>
-            This chart shows an estimate of contribution activity. Exact daily counts are not available for public profiles.
-          </p>
-        </div>
-      )}
+      <HeatmapWithYearNav
+        username={user.login}
+        initialWeeks={heatmap}
+        initialCurrentStreak={currentStreak}
+        initialLongestStreak={longestStreak}
+      />
 
       {/* Back to top */}
       {showBackToTop && (
@@ -1327,247 +879,25 @@ export function ProfileView({
             borderRadius: "10px",
             cursor: "pointer",
             boxShadow: "0 2px 8px rgba(0, 0, 0, 0.15)",
-            transition: "transform 0.15s, background-color 0.15s",
             zIndex: 50,
           }}
-          onMouseEnter={(e) => {
-            e.currentTarget.style.transform = "translateY(-2px)";
-            e.currentTarget.style.backgroundColor = "#36b97e";
-          }}
-          onMouseLeave={(e) => {
-            e.currentTarget.style.transform = "translateY(0)";
-            e.currentTarget.style.backgroundColor = "#3ecf8e";
-          }}
         >
-          <svg
-            width="20"
-            height="20"
-            viewBox="0 0 24 24"
-            fill="none"
-            stroke="currentColor"
-            strokeWidth="2.5"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-            aria-hidden="true"
-          >
+          <svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.5" strokeLinecap="round" strokeLinejoin="round" aria-hidden="true">
             <polyline points="18 15 12 9 6 15" />
           </svg>
         </button>
       )}
 
-      {/* Hidden profile card for download */}
-      <div style={{ position: "fixed", left: "-9999px", top: "-9999px", overflow: "hidden", pointerEvents: "none" }}>
-        <div
-          ref={cardRef}
-          style={{
-            width: "600px",
-            height: "300px",
-            padding: "32px",
-            backgroundColor: "#1c1c1c",
-            border: "1px solid rgba(255, 255, 255, 0.1)",
-            borderRadius: "12px",
-            display: "flex",
-            flexDirection: "column",
-            justifyContent: "space-between",
-            boxSizing: "border-box",
-            fontFamily: "system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif",
-          }}
-        >
-          <div style={{ display: "flex", justifyContent: "space-between", alignItems: "flex-start", gap: "24px" }}>
-            {/* Left section: user & score */}
-            <div style={{ display: "flex", flexDirection: "column", gap: "24px" }}>
-              {/* User */}
-              <div style={{ display: "flex", gap: "16px", alignItems: "center" }}>
-                {/* eslint-disable-next-line @next/next/no-img-element */}
-                <img
-                  src={user.avatar_url}
-                  alt={displayName}
-                  style={{ width: "64px", height: "64px", borderRadius: "9999px", border: "1px solid rgba(255, 255, 255, 0.15)", objectFit: "cover" }}
-                  crossOrigin="anonymous"
-                />
-                <div>
-                  <div style={{ fontSize: "18px", fontWeight: 600, color: "#ffffff", letterSpacing: "-0.3px", lineHeight: 1.2 }}>
-                    {displayName}
-                  </div>
-                  <div style={{ fontSize: "13px", color: "#9a9a9a", marginTop: "2px" }}>
-                    @{user.login}
-                  </div>
-                </div>
-              </div>
-              {/* Score */}
-              <div>
-                <div style={{ fontSize: "10px", textTransform: "uppercase", letterSpacing: "1px", color: "#9a9a9a", fontWeight: 600 }}>
-                  Contributor Score
-                </div>
-                <div style={{ fontSize: "44px", fontWeight: 700, color: "#3ecf8e", marginTop: "4px", lineHeight: 1 }}>
-                  {score}
-                </div>
-              </div>
-            </div>
-
-            {/* Right section: stats grid */}
-            <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr", gap: "12px", width: "260px" }}>
-              {[
-                { label: "Commits", value: stats.totalCommits },
-                { label: "PRs", value: stats.totalPRs },
-                { label: "Issues", value: stats.totalIssues },
-                { label: "Reviews", value: stats.totalReviews },
-              ].map((stat) => (
-                <div
-                  key={stat.label}
-                  style={{
-                    border: "1px solid rgba(255, 255, 255, 0.08)",
-                    borderRadius: "8px",
-                    padding: "12px 14px",
-                    backgroundColor: "#202020",
-                    display: "flex",
-                    flexDirection: "column",
-                    justifyContent: "center",
-                  }}
-                >
-                  <div style={{ fontSize: "20px", fontWeight: 600, color: "#ffffff", lineHeight: 1.1 }}>
-                    {stat.value.toLocaleString("en-US")}
-                  </div>
-                  <div style={{ fontSize: "11px", color: "#9a9a9a", marginTop: "4px", fontWeight: 500 }}>
-                    {stat.label}
-                  </div>
-                </div>
-              ))}
-            </div>
-          </div>
-
-          {/* Footer branding */}
-          <div
-            style={{
-              display: "flex",
-              justifyContent: "space-between",
-              alignItems: "center",
-              borderTop: "1px solid rgba(255, 255, 255, 0.08)",
-              paddingTop: "16px",
-              marginTop: "16px",
-            }}
-          >
-            <div style={{ display: "flex", alignItems: "center", gap: "6px" }}>
-              <span style={{ width: "8px", height: "8px", borderRadius: "50%", backgroundColor: "#3ecf8e" }} />
-              <span style={{ fontSize: "14px", fontWeight: 600, color: "#ffffff", letterSpacing: "-0.2px" }}>OSSfolio</span>
-            </div>
-            <span style={{ fontSize: "11px", fontFamily: "ui-monospace, Menlo, Monaco, Consolas, monospace", color: "#707070" }}>
-              ossfolio.qzz.io
-            </span>
-          </div>
-        </div>
-      </div>
-
-      {/* Add Badge Modal */}
+      {/* Badge Modal */}
       {isOwner && (
-        <dialog
-          ref={dialogRef}
-          style={{
-            position: "fixed",
-            top: "50%",
-            left: "50%",
-            transform: "translate(-50%, -50%)",
-            margin: 0,
-            border: "1px solid var(--color-hairline)",
-            borderRadius: "12px",
-            padding: "24px",
-            backgroundColor: "var(--color-canvas)",
-            color: "var(--color-ink)",
-            maxWidth: "400px",
-            width: "calc(100% - 32px)",
-            boxShadow: "0 10px 25px -5px rgba(0, 0, 0, 0.1), 0 8px 10px -6px rgba(0, 0, 0, 0.1)",
-          }}
-        >
-          <div style={{ display: "flex", flexDirection: "column", gap: "16px" }}>
-            <h3 style={{ fontSize: "18px", fontWeight: 600, margin: 0, color: "var(--color-ink)", letterSpacing: "-0.2px" }}>
-              Add Program Badge
-            </h3>
-
-            <div style={{ display: "flex", flexDirection: "column", gap: "6px" }}>
-              <label htmlFor="program-select" style={{ fontSize: "13px", fontWeight: 500, color: "var(--color-ink-mute)" }}>
-                Program
-              </label>
-              <select
-                id="program-select"
-                value={selectedProgram}
-                onChange={(e) => setSelectedProgram(e.target.value)}
-                style={{
-                  padding: "8px 12px",
-                  borderRadius: "6px",
-                  border: "1px solid var(--color-hairline-strong)",
-                  backgroundColor: "var(--color-canvas)",
-                  color: "var(--color-ink)",
-                  fontSize: "14px",
-                  width: "100%",
-                }}
-              >
-                {["GSSoC", "Hacktoberfest", "EluSoC", "GSoC", "MLH Fellowship", "SWoC"].map((p) => (
-                  <option key={p} value={p}>{p}</option>
-                ))}
-              </select>
-            </div>
-
-            <div style={{ display: "flex", flexDirection: "column", gap: "6px" }}>
-              <label htmlFor="year-select" style={{ fontSize: "13px", fontWeight: 500, color: "var(--color-ink-mute)" }}>
-                Year
-              </label>
-              <select
-                id="year-select"
-                value={selectedYear}
-                onChange={(e) => setSelectedYear(Number(e.target.value))}
-                style={{
-                  padding: "8px 12px",
-                  borderRadius: "6px",
-                  border: "1px solid var(--color-hairline-strong)",
-                  backgroundColor: "var(--color-canvas)",
-                  color: "var(--color-ink)",
-                  fontSize: "14px",
-                  width: "100%",
-                }}
-              >
-                {Array.from({ length: currentYear - 2000 + 1 }, (_, i) => currentYear - i).map((y) => (
-                  <option key={y} value={y}>{y}</option>
-                ))}
-              </select>
-            </div>
-
-            <div style={{ display: "flex", justifyContent: "flex-end", gap: "8px", marginTop: "8px" }}>
-              <button
-                type="button"
-                onClick={() => setIsBadgeModalOpen(false)}
-                style={{
-                  padding: "8px 14px",
-                  fontSize: "13px",
-                  fontWeight: 500,
-                  color: "var(--color-ink)",
-                  backgroundColor: "var(--color-canvas)",
-                  border: "1px solid var(--color-hairline-strong)",
-                  borderRadius: "6px",
-                  cursor: "pointer",
-                }}
-              >
-                Cancel
-              </button>
-              <button
-                type="button"
-                onClick={handleAddBadge}
-                disabled={isSavingBadge}
-                style={{
-                  padding: "8px 14px",
-                  fontSize: "13px",
-                  fontWeight: 500,
-                  color: "#ffffff",
-                  backgroundColor: "#3ecf8e",
-                  border: "none",
-                  borderRadius: "6px",
-                  cursor: isSavingBadge ? "not-allowed" : "pointer",
-                }}
-              >
-                {isSavingBadge ? "Saving..." : "Add"}
-              </button>
-            </div>
-          </div>
-        </dialog>
+        <ProfileBadgeModal
+          open={isBadgeModalOpen}
+          onClose={() => setIsBadgeModalOpen(false)}
+          badgesList={badgesList}
+          onBadgesUpdate={setBadgesList}
+          profileId={profileId}
+          username={user.login}
+        />
       )}
     </div>
   );
